feat(orm,domain): bigint RowVersion trigger replaces xmin concurrency token
xmin would have been reset by VACUUM FREEZE when its age crossed
vacuum_freeze_table_age (default 150M txns). On a high-volume table
that event invalidates every cached ETag at once — every client that
read a row before the freeze gets a wave of 412 Precondition Failed
on its next If-Match write.

Switch to a dedicated 'RowVersion' bigint column on Sales, maintained
by a BEFORE UPDATE trigger (ambev_sales_bump_rowversion) that does
NEW.RowVersion := OLD.RowVersion + 1. EF Core treats it the same way
(IsConcurrencyToken + ValueGeneratedOnAddOrUpdate), so the wire-level
flow is unchanged: WHERE RowVersion = @old in the UPDATE, RETURNING
the new value. The token is monotonic across the row's entire life
and is unaffected by FREEZE.

CLR types follow: Sale.RowVersion is now long (was uint), SaleDto and
the Update/Delete commands' ExpectedRowVersion are long?, the
controller's ETagFor / ParseIfMatch helpers parse a hex long. The
ETag wire format stays hex so existing clients that just round-trip
the value keep working — only its width grows.

+++ template/backend/src/Ambev.DeveloperEvaluation.Application/Sales/Common/SaleDto.cs
@@ -16,6 +16,6 @@ public class SaleDto
     public bool IsCancelled { get; set; }
     public DateTime CreatedAt { get; set; }
     public DateTime? UpdatedAt { get; set; }
-    public uint RowVersion { get; set; }
+    public long RowVersion { get; set; }
     public List<SaleItemDto> Items { get; set; } = new();
 }
